Update MBA_insights.py
Imported necessary libraries and the orders and association rules data sets, which were put into data frames. Dashboard titles and sidebar titles were created and 2 tabs were initialized to show the visualizations of the 2 datasets separately.  The start and end date and  product category filters were added and the logic was initialized to make the filters change the dashboard.